requirements.txt
aiohttp==3.8.1
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
frozenlist==1.3.3
idna==3.4
multidict==6.0.2
psycopg2-binary==2.9.3
pyTelegramBotAPI==4.6.0
python-dotenv==0.21.0
pytz==2022.1
requests==2.28.1
Shapely==1.8.5.post1
tzwhere==3.0.3
urllib3==1.26.12
yarl==1.8.1
